Rename data.rs to block_data.rs
